backend/requirements.txt
alembic==1.14.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.5.2
backports.zoneinfo==0.2.1
bcrypt==4.3.0
billiard==4.2.1
celery==5.5.2
cffi==1.17.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.2
dnspython==2.6.1
ecdsa==0.19.1
email-validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.12
greenlet==3.1.1
h11==0.14.0
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
jinja2==3.1.6
kombu==5.5.3
mako==1.3.10
Markdown==3.7
MarkupSafe==2.1.5
passlib==1.7.4
prompt-toolkit==3.0.51
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
pydantic-settings==2.8.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
python-jose==3.4.0
python-slugify==8.0.4
pytz==2025.2
rsa==4.9.1
sendgrid==6.11.0
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.40
starkbank-ecdsa==2.2.0
starlette==0.44.0
text-unidecode==1.3
typing-extensions==4.13.2
tzdata==2025.2
uvicorn==0.33.0
vine==5.1.0
wcwidth==0.2.13
zipp==3.20.2
